Fixing problem with loading a relative path.
If your path had a ".." or "." in it anywhere it would not fix the
schema name (id) properly. This was because of the call to path.resolve
which would fix all the ".." and ".", however, relativeTo would be
passed as the original value with the dots. When it tried to do the
string replace it would fail to find anything to replace.